# docs/swagger.yaml
basePath: /v1
definitions:
  contract.Metadata:
    properties:
      current_page:
        type: integer
      first_page:
        type: integer
      last_page:
        type: integer
      page_size:
        type: integer
      total_records:
        type: integer
    type: object
...
  contract.ReqOrderPay:
    properties:
      game_id:
        type: integer
      payment_method:
        enum:
        - ALFAMART
        - INDOMARET
        - OVO
        - SALDO
        type: string
    required:
    - payment_method
    type: object
...
  contract.ResCartUpdate:
    properties:
      cart:
        properties:
          cart_id:
            type: integer
          game_id:
            type: integer
          quantity:
            type: integer
          sub_total:
            type: number
        type: object
      message:
        type: string
    type: object
  contract.ResGameGetAll:
    properties:
      games:
        items:
          properties:
            description:
              type: string
            game_id:
              type: integer
            genres:
              items:
                type: string
              type: array
            price:
              type: number
            stock:
              type: integer
            title:
              type: string
            year:
              type: integer
          type: object
        type: array
      metadata:
        $ref: '#/definitions/contract.Metadata'
    type: object
  contract.ResGameGetByID:
    properties:
      game:
        properties:
          description:
            type: string
          game_id:
            type: integer
          genres:
            items:
              type: string
            type: array
          price:
            type: number
          stock:
            type: integer
          title:
            type: string
          year:
            type: integer
        type: object
    type: object
  contract.ResOrderPay:
    properties:
      order:
        properties:
          created_at:
            type: string
          order_details:
            items:
              properties:
                game:
                  properties:
                    description:
                      type: string
                    game_id:
                      type: integer
                    genres:
                      items:
                        type: string
                      type: array
                    price:
                      type: number
                    title:
                      type: string
                    year:
                      type: integer
                  type: object
                quantity:
                  type: integer
                sub_total:
                  type: number
              type: object
            type: array
          order_id:
            type: integer
          payment:
            properties:
              invoice_url:
                type: string
              payment_method:
                type: string
              status:
                type: string
            type: object
          status:
            type: string
          total_price:
            type: number
          updated_at:
            type: string
          user:
            properties:
              address:
                type: string
              email:
                type: string
              first_name:
                type: string
              last_name:
                type: string
              phone:
                type: string
              user_id:
                type: integer
            required:
            - address
            - phone
            type: object
        type: object
    type: object
...
host: localhost:8080
info:
  contact:
    email: ucokkocu411@gmail.com
    name: ucok-man
  description: Documentation for PixelRental API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PixelRental API
  version: "1.0"
paths:
...
  /games:
    get:
      consumes:
      - application/json
      description: Get all available games record
      parameters:
      - description: game title
        in: query
        name: title
        type: string
      - description: game genres
        in: query
        name: genres
        type: string
      - description: sorting param
        in: query
        name: sort
        type: string
      - description: current page
        in: query
        name: page
        type: string
      - description: page size param
        in: query
        name: page_size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contract.ResGameGetAll'
        "422":
          description: Unprocessable Entity
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "429":
          description: Too Many Requests
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
      summary: Get all games
      tags:
      - games
  /games/:id:
    get:
      consumes:
      - application/json
      description: Getting info of games record
      parameters:
      - description: games id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contract.ResGameGetByID'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "429":
          description: Too Many Requests
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
      summary: Get games by id
      tags:
      - games
  /orders:
    get:
      consumes:
      - application/json
      description: Get all available order record
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contract.ResCartUpdate'
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "429":
          description: Too Many Requests
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
      summary: Get all order
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create new order record
      parameters:
      - description: Create Order
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/contract.ReqOrderPay'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/contract.ResOrderPay'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "429":
          description: Too Many Requests
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
      summary: Create order
      tags:
      - orders
...
swagger: "2.0"